Resolve conflicts with nullable decimal support from main while keeping nullable DateTime and DateTimeOffset RPC support in the server-extension lowering and runtime marshaller paths.

Update negative nullable tests to use short?, because decimal? is now supported on main, and keep the required-test budget entries for both decimal and DateTime coverage.

+++ tests/DotBoxD.Kernels.Tests/Plugins/Rpc/Kernel/KernelRpcClientProxyValidationTests.cs
@@ -146,13 +146,13 @@ public sealed class ServerExtensionClientProxyValidationTests
 
             public interface IEchoService
             {
-                ValueTask<int> EchoAsync(decimal? value);
+                ValueTask<int> EchoAsync(short? value);
             }
 
             [ServerExtension("echo", typeof(IEchoService))]
             public sealed partial class EchoKernel
             {
-                public int Echo(decimal? value, HookContext ctx)
+                public int Echo(short? value, HookContext ctx)
                 {
                     return 0;
                 }
